add error handling and fix return issue for POST
ecspcs-178

This commit adds error handling but tackles another issue.
Because the new platform comes in missing a few fields returning what comes in results in null for
fields that aren't actually null. To get around this I look up the newly created object by name and
grab it from the database. This way the user sees a full object w/id and date fields filled