implement a near complete GET for games
ecspcs-186

This commit introduces a GET for games as well as the necessary files for this operation to work.
I wanted to introduce a GET with out the added complexity of a list of Platforms the game is compatible
with. Hence the 'near complete.' This way I could verify that other pieces are working before messing
around with the extra layer.